variable type error fixed in udp packet

+++ Sidus_Github_Solution/Drone_Flight_Controller_2/cMsgUdpR01.h
@@ -24,9 +24,9 @@ struct structMsgUdpR01
 	short mpuAccZ;
 
 	//Accelerometer Measurements without Gravity
-	short mpuAccWorldX;
-	short mpuAccWorldY;
-	short mpuAccWorldZ;
+	float mpuAccWorldX;
+	float mpuAccWorldY;
+	float mpuAccWorldZ;
 
 	//Yaw Pitch Roll Measurements in Radians
 	float mpuYaw;
